Backend/micro_services/SERVICE_AUTHENTIFICATION/service_utilisateur/requirements.txt
Flask==2.3.0
Flask-CORS==4.0.0
PyMySQL==1.1.0
bcrypt==4.1.2
PyJWT==2.10.1
